Sửa database lần thứ n

+++ DAL/CHITIETHOADON.cs
@@ -23,8 +23,8 @@ namespace DAL
 
         public int? SOLUONG { get; set; }
 
-        [Column(TypeName = "money")]
-        public decimal? THANHTIEN { get; set; }
+        [Column(TypeName = "int")]
+        public int? THANHTIEN { get; set; }
 
         public virtual HOADON HOADON { get; set; }
